// src/app/Core/Enums/SocialMediaPlatform.ts
// Enum for social media platforms
export enum SocialMediaPlatform {
    Facebook = 'Facebook',
    Instagram = 'Instagram',
    LinkedIn = 'LinkedIn',
    Twitter = 'Twitter',
    YouTube = 'YouTube',
    Pinterest = 'Pinterest',
    Snapchat = 'Snapchat',
    TikTok = 'TikTok',
    WhatsApp = 'WhatsApp',
    Reddit = 'Reddit'
}  